+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/connector/compressor/PipeSnappyCompressor.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.commons.pipe.connector.compress;
+package org.apache.iotdb.commons.pipe.connector.compressor;
 
 import org.apache.tsfile.compress.ICompressor;
 import org.apache.tsfile.compress.IUnCompressor;
@@ -25,13 +25,13 @@ import org.apache.tsfile.file.metadata.enums.CompressionType;
 
 import java.io.IOException;
 
-public class PipeLZMA2Compressor extends PipeCompressor {
+public class PipeSnappyCompressor extends PipeCompressor {
 
-  ICompressor compressor = ICompressor.getCompressor(CompressionType.LZMA2);
-  IUnCompressor decompressor = IUnCompressor.getUnCompressor(CompressionType.LZMA2);
+  private final ICompressor compressor = ICompressor.getCompressor(CompressionType.SNAPPY);
+  private final IUnCompressor decompressor = IUnCompressor.getUnCompressor(CompressionType.SNAPPY);
 
-  public PipeLZMA2Compressor() {
-    compressionType = PipeCompressionType.LZMA2;
+  public PipeSnappyCompressor() {
+    super(PipeCompressionType.SNAPPY);
   }
 
   @Override
